projects/dataflow-framework/abstraction-level-7/pyproject.toml
[project]
name = "abstraction-level-7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.6",
    "networkx>=3.5",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "typer>=0.17.3",
]
